changed my params from int8 and int16 to int32 to be compatible with mavsdk

+++ ArduCopter/Copter.cpp
@@ -235,7 +235,7 @@ void Copter::setup()
 
 void Copter::set_and_save_watts_parameters()
 {
-    int16_t group_id = g.prop_grp_id;
+    int32_t group_id = g.prop_grp_id;
     switch (group_id) {
     case 0: // invalid
         break;
